+++ server/src/main/java/org/opensearch/search/backpressure/SearchBackpressureService.java
@@ -16,12 +16,16 @@ import org.opensearch.common.component.AbstractLifecycleComponent;
 import org.opensearch.common.util.TokenBucket;
 import org.opensearch.monitor.jvm.JvmStats;
 import org.opensearch.monitor.process.ProcessProbe;
+import org.opensearch.search.backpressure.settings.SearchBackpressureMode;
 import org.opensearch.search.backpressure.settings.SearchBackpressureSettings;
+import org.opensearch.search.backpressure.stats.SearchBackpressureStats;
+import org.opensearch.search.backpressure.stats.SearchShardTaskStats;
 import org.opensearch.search.backpressure.trackers.CpuUsageTracker;
 import org.opensearch.search.backpressure.trackers.ElapsedTimeTracker;
 import org.opensearch.search.backpressure.trackers.HeapUsageTracker;
 import org.opensearch.search.backpressure.trackers.NodeDuressTracker;
 import org.opensearch.search.backpressure.trackers.TaskResourceUsageTracker;
+import org.opensearch.search.backpressure.trackers.TaskResourceUsageTrackerType;
 import org.opensearch.tasks.CancellableTask;
 import org.opensearch.tasks.Task;
 import org.opensearch.tasks.TaskCancellation;
@@ -117,7 +121,8 @@ public class SearchBackpressureService extends AbstractLifecycleComponent
     }
 
     void doRun() {
-        if (getSettings().isEnabled() == false) {
+        SearchBackpressureMode mode = getSettings().getMode();
+        if (mode == SearchBackpressureMode.DISABLED) {
             return;
         }
 
@@ -126,7 +131,7 @@ public class SearchBackpressureService extends AbstractLifecycleComponent
         }
 
         // We are only targeting in-flight cancellation of SearchShardTask for now.
-        List<CancellableTask> searchShardTasks = getSearchShardTasks();
+        List<SearchShardTask> searchShardTasks = getSearchShardTasks();
 
         // Force-refresh usage stats of these tasks before making a cancellation decision.
         taskResourceTrackingService.refreshResourceStats(searchShardTasks.toArray(new Task[0]));
@@ -138,12 +143,13 @@ public class SearchBackpressureService extends AbstractLifecycleComponent
 
         for (TaskCancellation taskCancellation : getTaskCancellations(searchShardTasks)) {
             logger.debug(
-                "cancelling task [{}] due to high resource consumption [{}]",
+                "[{} mode] cancelling task [{}] due to high resource consumption [{}]",
+                mode.getName(),
                 taskCancellation.getTask().getId(),
                 taskCancellation.getReasonString()
             );
 
-            if (getSettings().isEnforced() == false) {
+            if (mode != SearchBackpressureMode.ENFORCED) {
                 continue;
             }
 
@@ -159,7 +165,6 @@ public class SearchBackpressureService extends AbstractLifecycleComponent
             }
 
             taskCancellation.cancel();
-            state.incrementCancellationCount();
         }
     }
 
@@ -182,7 +187,7 @@ public class SearchBackpressureService extends AbstractLifecycleComponent
     /**
      * Returns true if the increase in heap usage is due to search requests.
      */
-    boolean isHeapUsageDominatedBySearch(List<CancellableTask> searchShardTasks) {
+    boolean isHeapUsageDominatedBySearch(List<SearchShardTask> searchShardTasks) {
         long usage = searchShardTasks.stream().mapToLong(task -> task.getTotalResourceStats().getMemoryInBytes()).sum();
         long threshold = getSettings().getSearchShardTaskSettings().getTotalHeapBytesThreshold();
         if (usage < threshold) {
@@ -196,12 +201,12 @@ public class SearchBackpressureService extends AbstractLifecycleComponent
     /**
      * Filters and returns the list of currently running SearchShardTasks.
      */
-    List<CancellableTask> getSearchShardTasks() {
+    List<SearchShardTask> getSearchShardTasks() {
         return taskResourceTrackingService.getResourceAwareTasks()
             .values()
             .stream()
             .filter(task -> task instanceof SearchShardTask)
-            .map(task -> (CancellableTask) task)
+            .map(task -> (SearchShardTask) task)
             .collect(Collectors.toUnmodifiableList());
     }
 
@@ -222,13 +227,17 @@ public class SearchBackpressureService extends AbstractLifecycleComponent
             }
         }
 
+        if (task instanceof SearchShardTask) {
+            callbacks.add(state::incrementCancellationCount);
+        }
+
         return new TaskCancellation(task, reasons, callbacks);
     }
 
     /**
      * Returns a list of TaskCancellations sorted by descending order of their cancellation scores.
      */
-    List<TaskCancellation> getTaskCancellations(List<CancellableTask> tasks) {
+    List<TaskCancellation> getTaskCancellations(List<? extends CancellableTask> tasks) {
         return tasks.stream()
             .map(this::getTaskCancellation)
             .filter(TaskCancellation::isEligibleForCancellation)
@@ -246,7 +255,7 @@ public class SearchBackpressureService extends AbstractLifecycleComponent
 
     @Override
     public void onTaskCompleted(Task task) {
-        if (getSettings().isEnabled() == false) {
+        if (getSettings().getMode() == SearchBackpressureMode.DISABLED) {
             return;
         }
 
@@ -310,4 +319,17 @@ public class SearchBackpressureService extends AbstractLifecycleComponent
 
     @Override
     protected void doClose() throws IOException {}
+
+    public SearchBackpressureStats nodeStats() {
+        List<SearchShardTask> searchShardTasks = getSearchShardTasks();
+
+        SearchShardTaskStats searchShardTaskStats = new SearchShardTaskStats(
+            state.getCancellationCount(),
+            state.getLimitReachedCount(),
+            taskResourceUsageTrackers.stream()
+                .collect(Collectors.toUnmodifiableMap(t -> TaskResourceUsageTrackerType.fromName(t.name()), t -> t.stats(searchShardTasks)))
+        );
+
+        return new SearchBackpressureStats(searchShardTaskStats, getSettings().getMode());
+    }
 }